Move tests to dedicated file
See https://github.com/element-hq/synapse/pull/17512#discussion_r1726849798
